Excluindo set e só deixando map
